remove unused pages, updated config
