+++ components/common/LoadingSpinner.vue
@@ -1,0 +1,6 @@
+<template>
+    <div class="relative w-12 h-12">
+      <div class="absolute w-12 h-12 border-4 border-t-blue-500 border-r-transparent border-b-blue-500 border-l-transparent rounded-full animate-spin"></div>
+      <div class="absolute w-12 h-12 border-4 border-t-transparent border-r-blue-300 border-b-transparent border-l-blue-300 rounded-full animate-[spin_1s_linear_infinite_reverse]"></div>
+    </div>
+  </template>
