+++ nuxt.config.ts
@@ -3,7 +3,7 @@ export default defineNuxtConfig({
 
   runtimeConfig: {
     public: {
-      apiBaseUrl: process.env.NUXT_PUBLIC_API_BASE_URL
+      apiBaseUrl: "http://localhost:8000/api/v1"
     }
   },
